[refactor] LineGenerator 패키지 분리
